+++ src/core/functions/internal_functions/file/InternalModuleFile.ts
@@ -12,7 +12,6 @@ import {
     TFile,
     TFolder,
 } from "obsidian";
-import { UNSUPPORTED_MOBILE_TEMPLATE } from "utils/Constants";
 import { TemplaterError } from "utils/Error";
 import { ModuleName } from "editor/TpDocumentation";
 
@@ -56,7 +55,7 @@ export class InternalModuleFile extends InternalModule {
     }
 
     async generate_content(): Promise<string> {
-        return await this.app.vault.read(this.config.target_file);
+        return await app.vault.read(this.config.target_file);
     }
 
     generate_create_new(): (
@@ -64,7 +63,7 @@ export class InternalModuleFile extends InternalModule {
         filename: string,
         open_new: boolean,
         folder?: TFolder
-    ) => Promise<TFile> {
+    ) => Promise<TFile | undefined> {
         return async (
             template: TFile | string,
             filename: string,
@@ -109,9 +108,8 @@ export class InternalModuleFile extends InternalModule {
     }
 
     generate_cursor_append(): (content: string) => void {
-        return (content: string): string => {
-            const active_view =
-                this.app.workspace.getActiveViewOfType(MarkdownView);
+        return (content: string): string | undefined => {
+            const active_view = app.workspace.getActiveViewOfType(MarkdownView);
             if (active_view === null) {
                 log_error(
                     new TemplaterError(
@@ -128,26 +126,16 @@ export class InternalModuleFile extends InternalModule {
         };
     }
 
-    generate_exists(): (filename: string) => boolean {
-        return (filename: string) => {
-            // TODO: Remove this, only here to support the old way
-            let match;
-            if ((match = this.linkpath_regex.exec(filename)) !== null) {
-                filename = match[1];
-            }
-
-            const file = this.app.metadataCache.getFirstLinkpathDest(
-                filename,
-                ""
-            );
-            return file != null;
+    generate_exists(): (filename: string) => Promise<boolean> {
+        return async (filename: string) => {
+            return await app.vault.exists(filename);
         };
     }
 
-    generate_find_tfile(): (filename: string) => TFile {
+    generate_find_tfile(): (filename: string) => TFile | null {
         return (filename: string) => {
             const path = normalizePath(filename);
-            return this.app.metadataCache.getFirstLinkpathDest(path, "");
+            return app.metadataCache.getFirstLinkpathDest(path, "");
         };
     }
 
@@ -181,7 +169,7 @@ export class InternalModuleFile extends InternalModule {
             let inc_file_content: string;
 
             if (include_link instanceof TFile) {
-                inc_file_content = await this.app.vault.read(include_link);
+                inc_file_content = await app.vault.read(include_link);
             } else {
                 let match;
                 if ((match = this.linkpath_regex.exec(include_link)) === null) {
@@ -192,7 +180,7 @@ export class InternalModuleFile extends InternalModule {
                 }
                 const { path, subpath } = parseLinktext(match[1]);
 
-                const inc_file = this.app.metadataCache.getFirstLinkpathDest(
+                const inc_file = app.metadataCache.getFirstLinkpathDest(
                     path,
                     ""
                 );
@@ -202,10 +190,10 @@ export class InternalModuleFile extends InternalModule {
                         `File ${include_link} doesn't exist`
                     );
                 }
-                inc_file_content = await this.app.vault.read(inc_file);
+                inc_file_content = await app.vault.read(inc_file);
 
                 if (subpath) {
-                    const cache = this.app.metadataCache.getFileCache(inc_file);
+                    const cache = app.metadataCache.getFileCache(inc_file);
                     if (cache) {
                         const result = resolveSubpath(cache, subpath);
                         if (result) {
@@ -244,9 +232,7 @@ export class InternalModuleFile extends InternalModule {
     generate_move(): (path: string, file_to_move?: TFile) => Promise<string> {
         return async (path: string, file_to_move?: TFile) => {
             const file = file_to_move || this.config.target_file;
-            const new_path = normalizePath(
-                `${path}.${file.extension}`
-            );
+            const new_path = normalizePath(`${path}.${file.extension}`);
             const dirs = new_path.replace(/\\/g, "/").split("/");
             dirs.pop(); // remove basename
             if (dirs.length) {
@@ -255,26 +241,27 @@ export class InternalModuleFile extends InternalModule {
                     await window.app.vault.createFolder(dir);
                 }
             }
-            await this.app.fileManager.renameFile(
-                file,
-                new_path
-            );
+            await app.fileManager.renameFile(file, new_path);
             return "";
         };
     }
 
     generate_path(): (relative: boolean) => string {
         return (relative = false) => {
-            // TODO: Add mobile support
+            let vault_path = "";
             if (Platform.isMobileApp) {
-                return UNSUPPORTED_MOBILE_TEMPLATE;
+                const vault_adapter = app.vault.adapter.fs.uri;
+                const vault_base = app.vault.adapter.basePath;
+                vault_path = `${vault_adapter}/${vault_base}`;
+            } else {
+                if (app.vault.adapter instanceof FileSystemAdapter) {
+                    vault_path = app.vault.adapter.getBasePath();
+                } else {
+                    throw new TemplaterError(
+                        "app.vault is not a FileSystemAdapter instance"
+                    );
+                }
             }
-            if (!(this.app.vault.adapter instanceof FileSystemAdapter)) {
-                throw new TemplaterError(
-                    "app.vault is not a FileSystemAdapter instance"
-                );
-            }
-            const vault_path = this.app.vault.adapter.getBasePath();
 
             if (relative) {
                 return this.config.target_file.path;
@@ -294,18 +281,14 @@ export class InternalModuleFile extends InternalModule {
             const new_path = normalizePath(
                 `${this.config.target_file.parent.path}/${new_title}.${this.config.target_file.extension}`
             );
-            await this.app.fileManager.renameFile(
-                this.config.target_file,
-                new_path
-            );
+            await app.fileManager.renameFile(this.config.target_file, new_path);
             return "";
         };
     }
 
     generate_selection(): () => string {
         return () => {
-            const active_view =
-                this.app.workspace.getActiveViewOfType(MarkdownView);
+            const active_view = app.workspace.getActiveViewOfType(MarkdownView);
             if (active_view == null) {
                 throw new TemplaterError(
                     "Active view is null, can't read selection."
@@ -318,11 +301,13 @@ export class InternalModuleFile extends InternalModule {
     }
 
     // TODO: Turn this into a function
-    generate_tags(): string[] {
-        const cache = this.app.metadataCache.getFileCache(
-            this.config.target_file
-        );
-        return getAllTags(cache);
+    generate_tags(): string[] | null {
+        const cache = app.metadataCache.getFileCache(this.config.target_file);
+
+        if (cache) {
+            return getAllTags(cache);
+        }
+        return null;
     }
 
     // TODO: Turn this into a function