+++ src/core/functions/internal_functions/system/InternalModuleSystem.ts
@@ -17,7 +17,7 @@ export class InternalModuleSystem extends InternalModule {
 
     async create_dynamic_templates(): Promise<void> {}
 
-    generate_clipboard(): () => Promise<string> {
+    generate_clipboard(): () => Promise<string | null> {
         return async () => {
             // TODO: Add mobile support
             if (Platform.isMobileApp) {
@@ -32,15 +32,14 @@ export class InternalModuleSystem extends InternalModule {
         default_value: string,
         throw_on_cancel: boolean,
         multi_line: boolean
-    ) => Promise<string> {
+    ) => Promise<string | null> {
         return async (
             prompt_text: string,
             default_value: string,
             throw_on_cancel = false,
             multi_line = false
-        ): Promise<string> => {
+        ): Promise<string | null> => {
             const prompt = new PromptModal(
-                this.app,
                 prompt_text,
                 default_value,
                 multi_line
@@ -67,7 +66,7 @@ export class InternalModuleSystem extends InternalModule {
         items: T[],
         throw_on_cancel: boolean,
         placeholder: string,
-        limit?: number,
+        limit?: number
     ) => Promise<T> {
         return async <T>(
             text_items: string[] | ((item: T) => string),
@@ -77,11 +76,10 @@ export class InternalModuleSystem extends InternalModule {
             limit?: number
         ): Promise<T> => {
             const suggester = new SuggesterModal(
-                this.app,
                 text_items,
                 items,
                 placeholder,
-                limit,
+                limit
             );
             const promise = new Promise(
                 (
@@ -95,7 +93,7 @@ export class InternalModuleSystem extends InternalModule {
                 if (throw_on_cancel) {
                     throw error;
                 }
-                return null;
+                return null as T;
             }
         };
     }