+++ src/core/functions/user_functions/UserScriptFunctions.ts
@@ -1,21 +1,20 @@
-import { App, TFile } from "obsidian";
+import { TFile } from "obsidian";
 
 import TemplaterPlugin from "main";
 import { IGenerateObject } from "../IGenerateObject";
-import { RunningConfig } from "core/Templater";
 import { get_tfiles_from_folder } from "utils/Utils";
 import { errorWrapperSync, TemplaterError } from "utils/Error";
 
 export class UserScriptFunctions implements IGenerateObject {
-    constructor(private app: App, private plugin: TemplaterPlugin) {}
+    constructor(private plugin: TemplaterPlugin) {}
 
-    async generate_user_script_functions(
-    ): Promise<Map<string, Function>> {
-        const user_script_functions: Map<string, Function> = new Map();
+    async generate_user_script_functions(): Promise<
+        Map<string, () => unknown>
+    > {
+        const user_script_functions: Map<string, () => unknown> = new Map();
         const files = errorWrapperSync(
             () =>
                 get_tfiles_from_folder(
-                    this.app,
                     this.plugin.settings.user_scripts_folder
                 ),
             `Couldn't find user script folder "${this.plugin.settings.user_scripts_folder}"`
@@ -37,20 +36,24 @@ export class UserScriptFunctions implements IGenerateObject {
 
     async load_user_script_function(
         file: TFile,
-        user_script_functions: Map<string, Function>
+        user_script_functions: Map<string, () => unknown>
     ): Promise<void> {
-        let req = (s: string) => {
+        const req = (s: string) => {
             return window.require && window.require(s);
         };
-        let exp: Record<string, unknown> = {};
-        let mod = {
-            exports: exp
+        const exp: Record<string, unknown> = {};
+        const mod = {
+            exports: exp,
         };
 
-        const file_content = await this.app.vault.read(file);
-        const wrapping_fn = window.eval("(function anonymous(require, module, exports){" + file_content + "\n})");
+        const file_content = await app.vault.read(file);
+        const wrapping_fn = window.eval(
+            "(function anonymous(require, module, exports){" +
+                file_content +
+                "\n})"
+        );
         wrapping_fn(req, mod, exp);
-        const user_function = exp['default'] || mod.exports;
+        const user_function = exp["default"] || mod.exports;
 
         if (!user_function) {
             throw new TemplaterError(
@@ -62,11 +65,15 @@ export class UserScriptFunctions implements IGenerateObject {
                 `Failed to load user script ${file.path}. Default export is not a function.`
             );
         }
-        user_script_functions.set(`${file.basename}`, user_function);
+        user_script_functions.set(
+            `${file.basename}`,
+            user_function as () => unknown
+        );
     }
 
     async generate_object(): Promise<Record<string, unknown>> {
-        const user_script_functions = await this.generate_user_script_functions();
+        const user_script_functions =
+            await this.generate_user_script_functions();
         return Object.fromEntries(user_script_functions);
     }
 }