+++ src/handlers/EventHandler.ts
@@ -2,7 +2,6 @@ import TemplaterPlugin from "main";
 import { Templater } from "core/Templater";
 import { Settings } from "settings/Settings";
 import {
-    App,
     EventRef,
     Menu,
     MenuItem,
@@ -13,17 +12,16 @@ import {
 
 export default class EventHandler {
     private syntax_highlighting_event: EventRef;
-    private trigger_on_file_creation_event: EventRef;
+    private trigger_on_file_creation_event: EventRef | undefined;
 
     constructor(
-        private app: App,
         private plugin: TemplaterPlugin,
         private templater: Templater,
         private settings: Settings
     ) {}
 
     setup(): void {
-        this.app.workspace.onLayoutReady(() => {
+        app.workspace.onLayoutReady(() => {
             this.update_trigger_file_on_creation();
         });
         this.update_syntax_highlighting();
@@ -32,21 +30,21 @@ export default class EventHandler {
 
     update_syntax_highlighting(): void {
         if (this.plugin.settings.syntax_highlighting) {
-            this.syntax_highlighting_event = this.app.workspace.on(
+            this.syntax_highlighting_event = app.workspace.on(
                 "codemirror",
                 (cm) => {
                     cm.setOption("mode", "templater");
                 }
             );
-            this.app.workspace.iterateCodeMirrors((cm) => {
+            app.workspace.iterateCodeMirrors((cm) => {
                 cm.setOption("mode", "templater");
             });
             this.plugin.registerEvent(this.syntax_highlighting_event);
         } else {
             if (this.syntax_highlighting_event) {
-                this.app.vault.offref(this.syntax_highlighting_event);
+                app.vault.offref(this.syntax_highlighting_event);
             }
-            this.app.workspace.iterateCodeMirrors((cm) => {
+            app.workspace.iterateCodeMirrors((cm) => {
                 cm.setOption("mode", "hypermd");
             });
         }
@@ -54,7 +52,7 @@ export default class EventHandler {
 
     update_trigger_file_on_creation(): void {
         if (this.settings.trigger_on_file_creation) {
-            this.trigger_on_file_creation_event = this.app.vault.on(
+            this.trigger_on_file_creation_event = app.vault.on(
                 "create",
                 (file: TAbstractFile) =>
                     Templater.on_file_creation(this.templater, file)
@@ -62,7 +60,7 @@ export default class EventHandler {
             this.plugin.registerEvent(this.trigger_on_file_creation_event);
         } else {
             if (this.trigger_on_file_creation_event) {
-                this.app.vault.offref(this.trigger_on_file_creation_event);
+                app.vault.offref(this.trigger_on_file_creation_event);
                 this.trigger_on_file_creation_event = undefined;
             }
         }
@@ -70,7 +68,7 @@ export default class EventHandler {
 
     update_file_menu(): void {
         this.plugin.registerEvent(
-            this.app.workspace.on("file-menu", (menu: Menu, file: TFile) => {
+            app.workspace.on("file-menu", (menu: Menu, file: TFile) => {
                 if (file instanceof TFolder) {
                     menu.addItem((item: MenuItem) => {
                         item.setTitle("Create new note from template")